rebase: replace partial with _components
