+++ src/app/(admin)/account/billing/_components/license-code-card.tsx
@@ -1,6 +1,6 @@
-import { Button } from "../ui/button";
-import { Card, CardContent, CardHeader, CardTitle } from "../ui/card";
-import { Input } from "../ui/input";
+import { Button } from "@/components/ui/button";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { Input } from "@/components/ui/input";
 
 export const LicenseCodeCard = () => {
   // TODO: License Code Form
