+++ src/app/(admin)/account/billing/loading.tsx
@@ -1,4 +1,4 @@
-import { EventSettingsSkeleton } from "@/components/skeletons/event-settings-skeleton";
+import { EventSettingsSkeleton } from "../../events/[eventId]/settings/_components/event-settings-skeleton";
 
 export default function Loading() {
   return <EventSettingsSkeleton />;
