+++ src/app/(admin)/account/settings/page.tsx
@@ -1,4 +1,3 @@
-import { EditAccountForm } from "@/components/forms/edit-account-form";
 import {
   Card,
   CardContent,
@@ -6,6 +5,7 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
+import { EditAccountForm } from "./_components/edit-account-form";
 
 export default function AccountSettingsPage() {
   return (
