+++ src/app/(admin)/account/usage/_components/usage-card.tsx
@@ -4,8 +4,8 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-} from "../ui/card";
-import { Progress } from "../ui/progress";
+} from "@/components/ui/card";
+import { Progress } from "@/components/ui/progress";
 
 interface UsageCardProps {
   title: string;
