+++ src/app/(admin)/events/[eventId]/photos/_components/control-bar.tsx
