+++ src/app/(admin)/events/[eventId]/photos/_components/event-action-buttons.tsx
