+++ src/app/(admin)/events/[eventId]/photos/_components/event-select-images.tsx
