+++ src/app/(admin)/events/[eventId]/photos/_components/no-event-images.tsx
