+++ src/app/(admin)/events/[eventId]/photos/page.tsx
@@ -1,10 +1,10 @@
-import { EditEventGallery } from "@/components/partials/event/edit-event-gallery";
-import { EventActionButtons } from "@/components/partials/event/event-action-buttons";
-import { NoEventImages } from "@/components/partials/event/no-event-images";
 import { paths } from "@/routes/paths";
 import { getServerAuthSession } from "@/server/auth";
 import { api } from "@/trpc/server";
 import { redirect } from "next/navigation";
+import { EventActionButtons } from "./_components/event-action-buttons";
+import { EditEventGallery } from "./_components/edit-event-gallery";
+import { NoEventImages } from "./_components/no-event-images";
 
 // TODO:
 // - replace root with redirect to overview
