+++ src/app/(admin)/events/[eventId]/settings/_components/delete-event-control.tsx
