+++ src/app/(admin)/events/[eventId]/settings/_components/event-setting.tsx
