+++ src/app/(admin)/events/[eventId]/settings/_components/event-settings-skeleton.tsx
@@ -5,17 +5,10 @@ import {
   CardFooter,
   CardHeader,
   CardTitle,
-} from "../ui/card";
+} from "@/components/ui/card";
 
 export const EventSettingsSkeleton = () => (
   <div className="animate-pulse space-y-5 pb-20 md:space-y-8">
-    <div className="flex flex-col justify-between gap-3 md:flex-row md:items-center md:gap-0">
-      <header className="space-y-1.5">
-        <div className="h-7 w-48 rounded-lg bg-muted/80" />
-        <div className="h-5 w-32 rounded-lg bg-muted/40" />
-        <div className="h-5 w-36 rounded-lg bg-muted/40" />
-      </header>
-    </div>
     <div className="space-y-5">
       <Card className="border-muted/30">
         <CardHeader className="border-b border-muted/50">
