+++ src/app/(auth)/sign-in/loading.tsx
@@ -1,4 +1,4 @@
-import { SigninSkeleton } from "@/components/skeletons/sigin-skeleton";
+import { SigninSkeleton } from "./_components/sigin-skeleton";
 
 export default function Loading() {
   return <SigninSkeleton />;
