+++ src/app/(client)/(event)/[eventId]/(root)/error.tsx
