+++ src/app/(client)/(event)/[eventId]/(root)/loading.tsx
