+++ src/app/(client)/(event)/[eventId]/(root)/page.tsx
