+++ src/app/(client)/(event)/[eventId]/gallery/_components/gallery-sidebar.tsx
