+++ src/app/(client)/(event)/[eventId]/gallery/_components/render-gallery-images.tsx
