+++ src/app/(client)/(event)/[eventId]/gallery/layout.tsx
