--- src/app/component/page.tsx
@@ -1,5 +1,0 @@
-import { EventsSleketon } from "@/components/skeletons/events-skeleton";
-
-export default function Component() {
-  return <EventsSleketon />;
-}
