+++ src/components/gallery/config.ts
@@ -1,0 +1,1 @@
+export const PLACEHOLDER_URL = "/images/placeholder.jpg" as const;
