+++ src/components/gallery/index.ts
@@ -1,0 +1,12 @@
+import {
+  type GalleryData,
+  type GalleryHandlers,
+  type GalleryOptions,
+  Gallery,
+} from "./gallery";
+
+export { type GalleryData, type GalleryHandlers, type GalleryOptions };
+
+export { Gallery };
+
+export default Gallery;
