+++ src/components/modals/gallery-dropzone.tsx
